+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- **added:** added `splice` and `spare_capacity_mut` to `BumpVec`
+
 ## 0.10.4 (2024-10-18)
 - **improved:** increased performance of `(Mut)BumpString::from_str_in`